Organize classes into suitable packages
