+++ src/main/java/derek/Ui.java
@@ -1,4 +1,9 @@
-import java.io.IOException;
+package derek;
+
+import derek.command.Command;
+import derek.task.Task;
+import derek.task.TaskList;
+
 import java.util.Random;
 import java.util.Scanner;
