+++ src/main/java/derek/command/Command.java
@@ -1,3 +1,7 @@
+package derek.command;
+
+import derek.IncorrectCommandException;
+
 /**
  * The Command class processes user input commands and determines the type of command
  * based on the input string. It also provides methods to check the validity of the command
@@ -10,7 +14,7 @@ public class Command {
         this.command = command;
     }
 
-    public void isConsent() throws IncorrectCommandException{
+    public void isConsent() throws IncorrectCommandException {
         if (!(this.command.equalsIgnoreCase("Y") || this.command.equalsIgnoreCase("N"))) {
             throw new IncorrectCommandException("Hate me just say! Answer my question RRRRAHHHH!!!");
         }
