+++ src/main/java/derek/command/CompleteCommand.java
@@ -1,3 +1,10 @@
+package derek.command;
+
+import derek.Storage;
+import derek.task.Task;
+import derek.task.TaskList;
+import derek.Ui;
+
 public class CompleteCommand extends Command {
 
     public CompleteCommand(String command) {
