+++ src/main/java/derek/command/DeadlineCommand.java
@@ -1,7 +1,14 @@
-public class EventCommand extends TaskCommand {
-    public EventCommand(String command) {
+package derek.command;
+
+import derek.*;
+import derek.task.Task;
+import derek.task.TaskList;
+
+public class DeadlineCommand extends TaskCommand {
+    public DeadlineCommand(String command) {
         super(command);
     }
+
     public void execute(Task task, Storage storage, Ui ui) {
         TaskList taskList = storage.getTaskList();
         taskList.add(task);
