+++ src/main/java/derek/command/IncompleteCommand.java
@@ -1,3 +1,10 @@
+package derek.command;
+
+import derek.Storage;
+import derek.task.Task;
+import derek.task.TaskList;
+import derek.Ui;
+
 public class IncompleteCommand extends Command {
 
     public IncompleteCommand(String command) {
