+++ src/main/java/derek/task/EventTask.java
@@ -1,3 +1,5 @@
+package derek.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
