+++ src/main/java/derek/task/Task.java
@@ -1,3 +1,5 @@
+package derek.task;
+
 /**
  * The Task class represents a general task with a name and a completion status.
  * It provides methods to mark the task as completed or incomplete, and factory methods
